renamed AssetType
