+++ TentakelsAttacking2/TentakelsAttacking2.vcxproj
@@ -186,7 +186,7 @@
     <ClInclude Include="UI\Scene\public\SceneManager.h" />
     <ClInclude Include="UI\Elements\Button\public\ClassicButton.h" />
     <ClInclude Include="Helper\public\AssetManager.h" />
-    <ClInclude Include="Helper\public\AssetType.hpp" />
+    <ClInclude Include="Helper\public\HAssetType.hpp" />
     <ClInclude Include="UI\Elements\Button\public\Button.h" />
     <ClInclude Include="UI\public\Focus.h" />
     <ClInclude Include="UI\public\Focusable.h" />
